fixed minor things and file names
